:recycle: restructuring codebase
